--- app/models/contract.rb
@@ -1,3 +1,0 @@
-class Contract < ApplicationRecord
-  attribute :duration, :interval
-end
