--- app/models/existing_contract.rb
@@ -1,3 +1,0 @@
-class ExistingContract < Contract
-  has_one :support_case, class_name: "Support::Case"
-end
