--- app/models/new_contract.rb
@@ -1,3 +1,0 @@
-class NewContract < Contract
-  has_one :support_case, class_name: "Support::Case"
-end
