+++ app/models/support/case.rb
@@ -17,8 +17,8 @@ module Support
 
     has_one :hub_transition, class_name: "Support::HubTransition", dependent: :destroy
 
-    belongs_to :existing_contract, class_name: "Contract", optional: true
-    belongs_to :new_contract, class_name: "Contract", optional: true
+    belongs_to :existing_contract, class_name: "Support::ExistingContract", optional: true
+    belongs_to :new_contract, class_name: "Support::NewContract", optional: true
 
     accepts_nested_attributes_for :hub_transition, allow_destroy: true, reject_if: :all_blank
