+++ app/models/support/contract.rb
@@ -1,0 +1,5 @@
+module Support
+  class Contract < ApplicationRecord
+    attribute :duration, :interval
+  end
+end
