+++ spec/models/support/existing_contract_spec.rb
@@ -1,5 +1,5 @@
 require "rails_helper"
 
-RSpec.describe ExistingContract, type: :model do
+RSpec.describe Support::ExistingContract, type: :model do
   it { is_expected.to have_one(:support_case) }
 end
