+++ spec/models/support/new_contract_spec.rb
@@ -1,5 +1,5 @@
 require "rails_helper"
 
-RSpec.describe NewContract, type: :model do
+RSpec.describe Support::NewContract, type: :model do
   it { is_expected.to have_one(:support_case) }
 end
